1. 패키지명 변경 2. 불필요 파일 제거
