+++ data/src/main/kotlin/jth/com/thetrive/data/db/dao/SampleDao.kt
@@ -1,0 +1,16 @@
+package jth.com.thetrive.data.db.dao
+
+import androidx.room.Dao
+import androidx.room.Insert
+import androidx.room.OnConflictStrategy
+import androidx.room.Query
+import jth.com.thetrive.data.model.SampleEntity
+
+@Dao
+interface SampleDao {
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    suspend fun insert(data: SampleEntity)
+
+    @Query("SELECT * FROM sample")
+    suspend fun getSamples(): List<SampleEntity>
+}
