+++ data/src/main/kotlin/jth/com/thetrive/data/di/ApiModule.kt
@@ -1,0 +1,53 @@
+package jth.com.thetrive.data.di
+
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import jth.com.thetrive.data.api.TheTriveService
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import java.util.concurrent.TimeUnit
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+internal object ApiModule {
+
+    private const val CONNECT_TIMEOUT = 15L
+    private const val WRITE_TIMEOUT = 15L
+    private const val READ_TIMEOUT = 15L
+    private const val BASE_URL = "https://api.github.com/"
+
+    @Provides
+    @Singleton
+    fun provideOkHttpClient(): OkHttpClient {
+        return OkHttpClient.Builder().apply {
+            connectTimeout(CONNECT_TIMEOUT, TimeUnit.SECONDS)
+            writeTimeout(WRITE_TIMEOUT, TimeUnit.SECONDS)
+            readTimeout(READ_TIMEOUT, TimeUnit.SECONDS)
+            retryOnConnectionFailure(true)
+            addInterceptor(HttpLoggingInterceptor().apply {
+                level = HttpLoggingInterceptor.Level.BODY
+            })
+        }.build()
+    }
+
+    @Provides
+    @Singleton
+    fun providesRetrofit(client: OkHttpClient): Retrofit {
+        return Retrofit.Builder()
+            .baseUrl(BASE_URL)
+            .addConverterFactory(GsonConverterFactory.create())
+            .client(client)
+            .build()
+    }
+
+    @Provides
+    @Singleton
+    fun providesService(retrofit: Retrofit): TheTriveService {
+        return retrofit.create(TheTriveService::class.java)
+    }
+}
