+++ data/src/main/kotlin/jth/com/thetrive/data/model/SampleDTO.kt
@@ -1,0 +1,3 @@
+package jth.com.thetrive.data.model
+
+data class SampleDTO(val name : String? = null)
