+++ data/src/main/kotlin/jth/com/thetrive/data/model/SampleEntity.kt
@@ -1,0 +1,11 @@
+package jth.com.thetrive.data.model
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity(tableName = "sample")
+data class SampleEntity(
+    @PrimaryKey(autoGenerate = true)
+    val id: Long,
+    val name: String
+)
