--- data/src/main/kotlin/recruiting_test_base/data/api/SampleService.kt
@@ -1,9 +1,0 @@
-package recruiting_test_base.data.api
-
-import recruiting_test_base.data.model.SampleDTO
-import retrofit2.http.GET
-
-interface SampleService {
-    @GET("sample")
-    suspend fun getSample() : List<SampleDTO>
-}
