--- data/src/main/kotlin/recruiting_test_base/data/model/SampleDTO.kt
@@ -1,3 +1,0 @@
-package recruiting_test_base.data.model
-
-data class SampleDTO(val name : String? = null)
