--- data/src/main/kotlin/recruiting_test_base/data/model/SampleEntity.kt
@@ -1,11 +1,0 @@
-package recruiting_test_base.data.model
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity(tableName = "sample")
-data class SampleEntity(
-    @PrimaryKey(autoGenerate = true)
-    val id: Long,
-    val name: String
-)
