+++ domain/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest
-    package="base.architecture.domain">
+    package="jth.com.thetrive.domain">
 </manifest>
