+++ domain/src/main/kotlin/jth/com/thetrive/domain/model/LocalSample.kt
@@ -1,0 +1,6 @@
+package jth.com.thetrive.domain.model
+
+data class LocalSample(
+    val id: Long,
+    val name: String
+)
