+++ domain/src/main/kotlin/jth/com/thetrive/domain/model/Sample.kt
@@ -1,0 +1,3 @@
+package jth.com.thetrive.domain.model
+
+data class Sample(val name : String?)
