--- domain/src/main/kotlin/recruiting_test_base/domain/model/LocalSample.kt
@@ -1,6 +1,0 @@
-package recruiting_test_base.domain.model
-
-data class LocalSample(
-    val id: Long,
-    val name: String
-)
