--- domain/src/main/kotlin/recruiting_test_base/domain/model/Sample.kt
@@ -1,3 +1,0 @@
-package recruiting_test_base.domain.model
-
-data class Sample(val name : String?)
