+++ settings.gradle
@@ -12,7 +12,7 @@ dependencyResolutionManagement {
         mavenCentral()
     }
 }
-rootProject.name = "base.architecture"
+rootProject.name = "jth.com.thetrive"
 include ':ui'
 include ':data'
 include ':domain'
