+++ ui/src/main/kotlin/jth/com/thetrive/ui/SampleApplication.kt
@@ -1,0 +1,11 @@
+package jth.com.thetrive.ui
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class SampleApplication : Application() {
+    override fun onCreate() {
+        super.onCreate()
+    }
+}
