+++ ui/src/main/kotlin/jth/com/thetrive/ui/viewmodels/BaseViewModel.kt
@@ -1,0 +1,5 @@
+import androidx.lifecycle.ViewModel
+
+open class BaseViewModel : ViewModel() {
+
+}
