+++ ui/src/main/kotlin/jth/com/thetrive/ui/views/BaseActivity.kt
@@ -1,0 +1,9 @@
+import androidx.annotation.LayoutRes
+import androidx.appcompat.app.AppCompatActivity
+
+abstract class BaseActivity : AppCompatActivity() {
+    @LayoutRes
+    abstract fun getLayoutResId(): Int
+    abstract fun initializeViewModel()
+    abstract fun initializeUiEvent()
+}
