+++ ui/src/main/kotlin/jth/com/thetrive/ui/views/SampleActivity.kt
@@ -1,0 +1,17 @@
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class SampleActivity : BaseActivity() {
+    override fun getLayoutResId(): Int {
+        TODO("Not yet implemented")
+    }
+
+    override fun initializeViewModel() {
+        TODO("Not yet implemented")
+    }
+
+    override fun initializeUiEvent() {
+        TODO("Not yet implemented")
+    }
+
+}
