--- ui/src/main/kotlin/recruiting_test_base/presentation/SampleApplication.kt
@@ -1,11 +1,0 @@
-package recruiting_test_base.presentation
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class SampleApplication : Application() {
-    override fun onCreate() {
-        super.onCreate()
-    }
-}
