--- ui/src/main/kotlin/recruiting_test_base/presentation/viewmodels/BaseViewModel.kt
@@ -1,6 +1,0 @@
-package recruiting_test_base.presentation.viewmodels
-
-import androidx.lifecycle.ViewModel
-
-open class BaseViewModel : ViewModel() {
-}
