--- ui/src/main/kotlin/recruiting_test_base/presentation/views/BaseActivity.kt
@@ -1,6 +1,0 @@
-package recruiting_test_base.presentation.views
-
-import androidx.appcompat.app.AppCompatActivity
-
-open class BaseActivity : AppCompatActivity()  {
-}
