--- ui/src/main/kotlin/recruiting_test_base/presentation/views/SampleActivity.kt
@@ -1,7 +1,0 @@
-package recruiting_test_base.presentation.views
-
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class SampleActivity : BaseActivity() {
-}
